Move application details content to specific components
