+++ apps/gamify/web-app/src/app/pages/admin/applications/applications.module.ts
@@ -6,11 +6,19 @@ import { ApplicationsRoutingModule } from './applications-routing.module';
 import { DetailsComponent } from './details/details.component';
 import { UiModule } from '../../../ui.module';
 import { BadgeModalComponent } from '../../../shared/modals/badge-modal/badge-modal.component';
+import { ApplicationDetailsBadgesComponent } from '../../../shared/components/application-details-badges/application-details-badges.component';
+import { ApplicationDetailsGeneralComponent } from '../../../shared/components/application-details-general/application-details-general.component';
+import { ApplicationDetailsPlayersComponent } from '../../../shared/components/application-details-players/application-details-players.component';
+import { ApplicationDetailsModeratorsComponent } from '../../../shared/components/application-details-moderators/application-details-moderators.component';
 
 @NgModule({
   declarations: [
     DetailsComponent,
     BadgeModalComponent,
+    ApplicationDetailsGeneralComponent,
+    ApplicationDetailsBadgesComponent,
+    ApplicationDetailsPlayersComponent,
+    ApplicationDetailsModeratorsComponent,
   ],
   imports: [
     CommonModule,
