--- apps/gamify/web-app/src/app/pages/admin/applications/details/details.component.css
@@ -1,3 +1,0 @@
-.input-width {
-    width: 400px;
-}
