+++ apps/gamify/web-app/src/app/pages/admin/applications/details/details.component.html
@@ -10,76 +10,10 @@
 <!-- Name, Type (disabled) Description,  -->
 <mat-tab-group>
     <mat-tab label="General"> 
-        <div style="display: flex; flex-direction: column;">
-            <h2 style="margin-top: 10px;">General Application Details</h2>
-
-            <mat-form-field class="input-width" appearance="fill">
-                <mat-label>Name</mat-label>
-                <input matInput placeholder="Very Cool Game" value="M4M Walking Challenge">
-            </mat-form-field>
-            
-            <mat-form-field class="input-width" appearance="fill">
-                <mat-label>Type</mat-label>
-                <input matInput placeholder="Very Cool Game" value="CHALLENGE" disabled>
-            </mat-form-field>
-
-            <mat-form-field class="input-width" appearance="fill">
-                <mat-label>Description</mat-label>
-                <textarea matInput placeholder="How to play the game?"></textarea>
-            </mat-form-field>
-
-            <mat-form-field class="input-width" appearance="fill">
-                <mat-label>External application url</mat-label>
-                <input matInput placeholder="https://mycoolgame.io" value="">
-            </mat-form-field>
-
-            <div style="margin-bottom: 10px;">
-                <button mat-raised-button color="primary">Save</button>
-                <button mat-raised-button style="margin-left: 10px;" routerLink="/admin/applications">Cancel</button>
-            </div>
-        </div>
+        <coders-application-details-general></coders-application-details-general>
     </mat-tab>
     <mat-tab label="Badges">
-        <div style="display: flex; flex-direction: column;">
-            <h2 style="margin-top: 10px;">Application Badges</h2>
-
-            <div>
-                <button mat-raised-button color="primary" (click)="openCreateBadgeModal()">Create Badge</button> 
-            </div>
-
-            <table mat-table [dataSource]="badgesDataSource" class="mat-elevation-z2" style="margin: 20px 0">
-                <ng-container matColumnDef="id">
-                    <th mat-header-cell *matHeaderCellDef> ID </th>
-                    <td mat-cell *matCellDef="let element"> {{element.id}} </td>
-                </ng-container>
-
-                <ng-container matColumnDef="name">
-                    <th mat-header-cell *matHeaderCellDef> Name </th>
-                    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-                </ng-container>
-              
-                <ng-container matColumnDef="tier">
-                    <th mat-header-cell *matHeaderCellDef> Tier </th>
-                    <td mat-cell *matCellDef="let element"> {{element.tier}} </td>
-                </ng-container>
-
-                <ng-container matColumnDef="repeatedlyObtainable">
-                    <th mat-header-cell *matHeaderCellDef> Repeatedly Obtainable</th>
-                    <td mat-cell *matCellDef="let element"> {{element.repeatedlyObtainable}} </td>
-                </ng-container>
-
-                <ng-container matColumnDef="actions">
-                    <th mat-header-cell *matHeaderCellDef> Actions </th>
-                    <td mat-cell *matCellDef="let element">
-                        <button mat-raised-button color="primary" (click)="openEditBadgeModal(element.id)">Edit</button>
-                        <button mat-raised-button color="warn" style="margin-left: 6px">Delete</button>
-                    </td>
-                </ng-container>
-              
-                <tr mat-header-row *matHeaderRowDef="badgesDisplayedColumns"></tr>
-                <tr mat-row *matRowDef="let row; columns: badgesDisplayedColumns;"></tr>
-              </table>
-        </div>
+        <coders-application-details-badges></coders-application-details-badges>
     </mat-tab>
     <mat-tab label="Players"> Content 3 </mat-tab>
     <mat-tab label="Moderators"> Content 4 </mat-tab>
