+++ apps/gamify/web-app/src/app/shared/components/application-details-general/application-details-general.component.scss
@@ -1,0 +1,3 @@
+.input-width {
+    width: 400px;
+}
